requirements.txt
streamlit
sentence-transformers
transformers
faiss-cpu
torch
pandas
numpy
sacrebleu
rouge-score
